# refactor!: remove easyjson dependency