--- a/Makefile
+++ b/Makefile
@@ -9,7 +9,7 @@ policy.wasm: $(SOURCE_FILES) go.mod go.sum types_easyjson.go
 		--rm \
 		-e GOFLAGS="-buildvcs=false" \
 		-v ${PWD}:/src \
-		-w /src tinygo/tinygo:0.27.0 \
+		-w /src tinygo/tinygo:0.28.1 \
 		tinygo build -o policy.wasm -target=wasi -no-debug .
 
 artifacthub-pkg.yml: metadata.yml go.mod