… go.sum
… hack/generate-easyjson.sh
… types.go
… types_easyjson.go
… vendor/github.com/go-openapi/strfmt/.golangci.yml
… vendor/github.com/go-openapi/strfmt/Makefile
… vendor/github.com/go-openapi/strfmt/types.go
… vendor/github.com/josharian/intern/README.md
… vendor/github.com/josharian/intern/intern.go
… vendor/github.com/josharian/intern/license.md
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/controller_revision.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/controller_revision_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/controller_revision_list_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_condition_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_list_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_spec.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_spec_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_status_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_update_strategy.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/daemon_set_update_strategy_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_condition_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_list_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_spec.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_spec_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_status_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_strategy.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/deployment_strategy_easyjson.go
… vendor/github.com/kubewarden/k8s-objects/api/apps/v1/replica_set_condition_easyjson.go