## Description

This PR removes easyjson dependency from the mix by refactoring the example/template policy.
Also, it bumps the build container to `tinygo:0.28.1`.

<!-- Please provide the link to the documentation related to your change, if applicable -->
<!-- [Documentation](https://<insert your url>) -->

## Test

Existing tests should pass

## Additional Information

### Size comparison

before:
```
λ du -h policy.wasm
1.3M    policy.wasm
```

after: 
```
λ du -h policy.wasm
480K    policy.wasm
```

### Benchmarks

`kwctl bench`  results

before:

```
validate_settings [75.8 thousand iterations in 5.07s with 100.0 samples]:
        elapsed [min mean max]: [50.58µs 66.77µs 78.08µs] (sample data: med = 66.64µs, var = 10.04ms², stdde
v = 3.17µs)
validate warming up for 3.00s
validate mean warm up execution time 68.90µs running 75.8 thousand iterations
validate [75.8 thousand iterations in 5.29s with 100.0 samples]:
        elapsed [min mean max]: [69.51µs 69.79µs 70.83µs] (sample data: med = 69.78µs, var = 34.73µs², stdde
v = 186.37ns)
```

after:

```
validate_settings [80.8 thousand iterations in 5.26s with 100.0 samples]:
        elapsed [min mean max]: [56.70µs 65.54µs 97.78µs] (sample data: med = 64.76µs, var = 17.64ms², stdde
v = 4.20µs)
validate warming up for 3.00s
validate mean warm up execution time 108.80µs running 50.5 thousand iterations
validate [50.5 thousand iterations in 5.60s with 100.0 samples]:
        elapsed [min mean max]: [94.64µs 111.06µs 166.29µs] (sample data: med = 110.76µs, var = 47.25ms², st
ddev = 6.87µs)
```
---

`kwctl bench --warm-up-time 1 --measurement-time 1 --num-resamples 2 --num-samples 2` results

before:

```
validate_settings [15.2 thousand iterations in 1.02s with 100.0 samples]:
        elapsed [min mean max]: [22.47µs 69.78µs 416.87µs] (sample data: med = 65.86µs, var = 1.35s², stddev
 = 36.78µs)
validate warming up for 1.00s
validate mean warm up execution time 68.86µs running 15.2 thousand iterations
validate [15.2 thousand iterations in 1.04s with 100.0 samples]:
        elapsed [min mean max]: [55.74µs 68.49µs 77.68µs] (sample data: med = 68.55µs, var = 4.75ms², stddev
 = 2.18µs)
```

after:

```
validate_settings [20.2 thousand iterations in 1.31s with 100.0 samples]:
        elapsed [min mean max]: [23.45µs 66.31µs 275.55µs] (sample data: med = 64.56µs, var = 504.99ms², std
dev = 22.47µs)
validate warming up for 1.00s
validate mean warm up execution time 113.65µs running 10.1 thousand iterations
validate [10.1 thousand iterations in 1.14s with 100.0 samples]:
        elapsed [min mean max]: [95.51µs 112.59µs 197.84µs] (sample data: med = 112.33µs, var = 188.63ms², s
tddev = 13.73µs)
```